feat: ErrorAlertService with standard error alerts
Moved error messages into an error alert service.
For now it's only used on the ChangePassword page, but it can be used anywhere to provide standard error messages.

Some of these errors are a bit rough, and as such, caution should be taken in using them - just an easy way to provide feedback to the user.