Implement profile picture management: Add options to take, pick, upload, and remove photos (#323)
* #257: Add camera screen

Enable taking, saving, and removing profile pictures. Implement camera functionality using the Expo Camera library.

* #257; Cache photo on login

* #257; Fix image problem

* Fix wrong field names

* Update bucket return data and photo cache

This commit removes the meta data that the frontend added to the base 64 images since the bucket service now returns images including the meta data. Additionally, the profile pictures are now cached on the student object instead of the user

* Fetch photo during user registration

Move photo fetching logic into the setStudentInfo function to ensure the photo is automatically retrieved and stored in AsyncStorage during login or registration. This commit also fixes issues with setting and displaying the photo correctly in the views

* Add back button and camera UI improvements

* Enhance image handling and upload functionality in user API

- Added support for fetching images as base64 from the bucket
- Updated uploadPhoto function to use expo-file-system for file handling
- Improved error handling for photo uploads and fetching
- Modified CameraScreen to integrate new upload and fetch logic

* Add correct file type handling

* Remove unused uri to blob

* Add camera roll option

* Fix failing test cases with mocks

Add mock data for expo-camera and mock implementation for useFocusEffect from React Navigation

* Add delete image functionality

---------

Co-authored-by: LucaPurcilly <lp@enablment.com>