Dev to Staging PBI 286 (#334)
PBI 286
Course now saves progress so you continue from where you left off when taking a course

-------------------------------------------------------------------------------------------------------------------------
PBI 303
We added a page option on the profile page "Certificados" that takes the user to their certificate page

We implemented an empty state to the certificate page for when the student doesn't have any certificates yes

On the certificate page we display certificates; we made styling for the certificate data arriving from the server as close to the figma as we could replicate, and added a search bar and category filter to browse in them.

We added a certificate popup, that is displayed when the student wants to see the full certificate in detail and want to download it (download functionality not implemented yet)¨

We Implemented an API service function to generate a certificate for a student on a given course, that can be integrated with course completion when it works. (9 unittests written for this)

We made a minor change to the display stack for the tabnavigator, making it accept a profilestack instead of simply the profilescreen, so the navbar is displayed on all the pages in the profile; editprofile, certificades etc.